# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Pocket Money API
  description: Backend API for managing chores, ledger entries, and settlements in family groups
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.pocket-money.example.com
    description: Production server
...
paths:
...
  /api/v1/groups/{id}/chores:
    get:
      tags:
        - Chores
      summary: List chores
      description: Returns all chores for a group
      operationId: listChores
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of chores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChoreResponse'
        '400':
          description: Invalid group ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a member of this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Chores
      summary: Create chore
      description: Creates a new chore for the group (head only)
      operationId: createChore
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChoreRequest'
      responses:
        '201':
          description: Chore created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreResponse'
        '400':
          description: Bad request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only group head can create chores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/chores/{id}:
    patch:
      tags:
        - Chores
      summary: Update chore
      description: Updates a chore (head only)
      operationId: updateChore
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Chore ID (UUID)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChoreRequest'
      responses:
        '200':
          description: Chore updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreResponse'
        '400':
          description: Bad request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only group head can update chores or cannot modify system chore
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chore not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Chores
      summary: Delete chore
      description: Soft deletes a chore (head only)
      operationId: deleteChore
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Chore ID (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Chore deleted successfully
        '400':
          description: Invalid chore ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only group head can delete chores or cannot delete system chore
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chore not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    CreateChoreRequest:
      type: object
      properties:
        name:
          type: string
          description: Chore name
        description:
          type: string
          nullable: true
          description: Chore description
        amount:
          type: number
          format: double
          minimum: 0.01
          description: Amount earned for completing the chore
      required:
        - name
        - amount
    UpdateChoreRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
          description: Chore name
        description:
          type: string
          nullable: true
          description: Chore description
        amount:
          type: number
          format: double
          nullable: true
          description: Amount earned for completing the chore
    ChoreResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Chore ID
        group_id:
          type: string
          format: uuid
          description: Group ID
        name:
          type: string
          description: Chore name
        description:
          type: string
          nullable: true
          description: Chore description
        amount:
          type: number
          format: double
          description: Amount earned for completing the chore
        is_system:
          type: boolean
          description: Whether this is a system-generated chore (e.g., Settlement)
        created_at:
          type: string
          format: date-time
          description: Chore creation timestamp
      required:
        - id
        - group_id
        - name
        - amount
        - is_system
        - created_at
...